CCodecBufferChannel: Support HAL decryption via C2EncryptionInfo
Enables Codec2 components to perform both decryption and decoding by
passing encrypted data and its associated metadata directly to the HAL.

When a component returns `SM_ENCRYPTED_WITH_KEY` secure mode,
CCodecBufferChannel executes ICrypto::getKeyHandle() instead of
ICrypto::decrypt() as the decryption is executed in the codec2 HAL.
The necessary information is carried to the codec2 HAL based on the
HAL advertisement. In addition, LinearInputBuffers is used instead
of EncryptedLinearInputBuffers because the secure input buffer is
not necessary and is passed to the HAL in C2Work::input::buffers.

Bug: 409171443
Test: play protected content video
Flag: android.media.codec.decrypt_and_decode_in_hal
Change-Id: I30b57ea7fc19e21fa92e5608e1c72c65c059ab36