Stable stream id for virtual camera
Each new input stream receives an ID that corresponding
to the number of streams previously added.

For example, the first stream gets ID 0, the second gets ID 1, and so
on.

This allow the virtual camera owner to identify for which stream
a request is being made and write the data with the corresponding size.

Bug: 430870772
Test: android.virtualdevice.cts.camera.VirtualCameraTest#virtualCamera_streamId_matchesCreationOrde
Flag: android.companion.virtualdevice.flags.virtual_camera_stable_stream_id
Change-Id: I2dd69e34584e85b923d22c453ee905aeb8c46720