Fix frames transferred cache in aaudio.
In aaudio, there is a monotonic counter to work around for HAL reporting
position in int32_t. The int32_t is reset every time when the stream is
stopped. This may cause the position doubled everytime when resuming,
which may finally lead to overflow after thousands times of stopping and
resuming. The fix is to only reset the int32_t counter when the mmap
buffer is newly created. When the stream starts, it should offset to
reflect the correct hardware read position.

Bug: 438663150
Bug: 444745496
Bug: 444528237
Test: atest AAudioTests
Test: start and stop stream multiple times
Flag: EXEMPT BUGFIX
Change-Id: I6159661dff011c4717a1c7a65fe3df5aa53163f8