Use write counter to calculate the drain time.
The clock model uses the mmap position reported from the HAL to convert
position to time. The position may jump after stop and resume. In that
case, there is an offset calculated from the first position report. This
is counted in the getFramesWritten(). But this offset should not be
counted for calculating the drain time.

Bug: 447453729
Test: stop, restart
Flag: EXEMPT BUGFIX
Change-Id: Iac3d3253d91a1e69d70ebe498a68fec77e2d4549