Always return hardware position as the HAL reports.
The HAL is supposed to report the MMAP position monotonically.
Currently, the framework has a machnism to seek to right index if
the HAL doesn't report the position monotonically. The position is
supposed to be where the DSP is reading. In this CL, it honors the
HAL's report but add debugging log if it is not monotonically
increasing.

Bug: 448251037
Test: atest AAudioTests
Test: start, stop, resume stream
Flag: EXEMPT BUGFIX
Change-Id: I1bab86bfd3bbbd49b3db294abfd7a6c615c53fb4