Copy data after getting mmap position when resuming from standby.
Before this CL, when resuming from standby, aaudio copy data to where
it is when stopping if there is any unprocessed data left. This will
not work correctly as aaudio will have to catch up to whatever mmap
position the HAL is reported to avoid conflict with the HAL. The read
write counter will be updated at this case.

It is also hard to move the unprocessed data to right in front of where
the DSP writing for the capture case.

In this CL, the unprocessed data is copied to memory and write to shared
mmap buffer when getting the first valid mmap position. For the capture
case, given the buffer size is small, it may not be needed to copy the
unprocessed data.

Bug: 448998423
Test: atest AAudioTests
Test: start, pause/stop, standby or not standby, resume
Flag: EXEMPT BUGFIX
Change-Id: Ia945f2e5489142dd98129c30211fad6558bab86b