libaudiohal@aidl: Add support for nested 'drain(EARLY_NOTIFY)' requests
It is now allowed to issue 'drain(EARLY_NOTIFY)' while draining
is ongoing. This scenario happens when running
'CtsMediaPlayerTestCases.testSetNextMediaPlayer'.

Since AIDL HAL stream state machine does not allow 'drain' command
when draining, 'StreamHalAidl' stashes these requests and issues
a drain as soon as it receives the event of clip end from the HAL.
This reflects the original intention of the nested 'drain' request
to drain out the clip which was slotted into the HAL while the current
clip was draining.

Flag: EXEMPT BUGFIX
Bug: 422674075
Test: atest audiotrack_tests
Test: atest --test-filter=".*testSetNextMediaPlayer" CtsMediaPlayerTestCases
Change-Id: I576a8c2796ec32c7fa7984dae2f42665ba1ac274