Copy unprocessed data to new location after pause offload stream.
For offload stream, when the client pauses the stream, it will stop
immediately. That may cause there are some data left in the shared mmap
buffer that is not yet played. When resuming the stream, the HAL may report
a position that is not continuous from the last reported position before
pausing. In that case, it is better the framework copy the data to where
the HAL is reading after resuming.

Bug: 450296391
Test: atest AAudioTests
Test: OboeTester, start, pause, resume
Flag: EXEMPT BUGFIX
Change-Id: I30e10532a8cae2d3ac6d8c7a83d132c17ed661f1