cameraserver: Add back checkOp for OP_CAMERA in notifyCameraOpening()
When this check isn't there, its possible for a client to open camera from the background and through a racy check in startCameraStreamingOps(since it depends on StatusTracker thread) to leak a frame to camera HAL and back to it.

Instead, we should block camera open if the
app doesn't have the valid state to pass checkOp during open itself.

Bug: 447599496

Test: Manual; Client starting can't open camera when its in the
      background

Flag: EXEMPT BUGFIX
Change-Id: I03e746e90cdfa273c0c403dad06668f6a8aed087