AudioPolicyManager: fix combo device warning from apm_extract_one_audio_device()
setVolumeIndexForGroup() should not call getDeviceForVolume()
with device combos that are artificially built by merging
the requested device with the device selected by the engine
for the attributes of the requested volume group.
Instead, check device belonging to device sets directly.

Also reorganize setVolumeIndexForGroup() and rename variables
to make the logic more clear and optimize the number of operations.

Test: manual call volume with various routes
Flag: EXEMPT BUGFIX
Change-Id: Ia80a592d19c4a311f14eccc4d82c69aa0f4653bc