Increase timeout to wait for the first frame
Apparently 3s is not enough to initialize the virtual camera pipeline
and provide a first frame by the VD owner, so allow for a sufficient
amount of time. It's ok if this fails on the Camera framework side,
though we shouldn't set strong limitations on the virtual camera side.

Bug: 450131373
Test: atest VirtualCameraTest
Flag: EXEMPT BUGFIX
Change-Id: Icf2a1259c8558360c5e767151cd230db458ac5a6