Add mirrorDisplay to screenrecord
The existing screen recording logic used mirrorLayerStack to capture the
display, which does not truly mirror a physical display if the user
decides to specify the display ID in the command line.

This change updates the screenrecord to use mirrorDisplay instead of
mirrorLayerStack when the physical display ID is set from the command
line. mirrorDisplay mirrors the display and is not affected by changes
in the display layer stack.

Bug: 281084746
Flag: EXEMPT BUGFIX
Test: adb shell screenrecord --display-id 4619827677550801152 /sdcard/ScreenRecordings/output.mp4
Test: adb shell screenrecord /sdcard/ScreenRecordings/output.mp4
Change-Id: Ic63241afc1c85991cf1328acadfaaee222354d5c