Use preflight PermissionChecker in notifyCameraOpening
AppOpsManager doesn't support device aware permission/appop checking, so
this will fail permission/appop checks by apps that have CAMERA
permission on a VirtualDevice context and not on the default device
context. PermissionChecker takes into consideration the
attributionSource (that includes the deviceId).

Do only a preflight check of the camera appop, not a start data delivery
as in ag/36612164.

Bug: 454614172
Bug: 456663437
Bug: 447599496
Bug: 456442544
Bug: 456510190
Test: manual using Vdm demo sample app
Test: fixes b/447599496
Test: atest face_*_camera_test
Test: atest VtsAidlCameraServiceTargetTest
Test: atest VirtualCameraTest
Test: atest MultiViewTest
Flag: EXEMPT BUGFIX
Change-Id: Ie756a43fa6296f49212b4e42fcfa229d84b91b94