For shared streams, prepare buffer for start before call service.
For shared streams, the shared mmap buffer will not be freed when
entering standby mode. In that case, it is safe to prepare buffer
for start before calling to service side. It is also necessary to
do so to avoid service side wrongly assume the client side starts
flowing.

Bug: 451182365
Test: atest AAudioTests
Flag: EXEMPT BUGFIX
Change-Id: I7493fa0a1e247386f7998237e43f4be3e011efdc