Use wp::fromExisting when creating weak pointers in DeathPipe
The current implementation of DeathPipe keeps a raw pointer to the
parent IBinder object, and implicitly creates a wp<IBinder> object from
the raw pointer as needed. This can sometimes lead to incorrect ref
counting, which may lead to crashes.

This CL replaces the implicit wp<IBinder> creation with an explicit weak
pointer that is passed to DeathPipe in the constructor. The parent
creates the weak pointer using wp::fromExisting which is the recommended
way to create weak pointers from a managed reference.

Bug: 442219172
Test: atest ACameraNdkVendorTest passes
Test: Killing FaceHAL while streaming camera behaves as expected.
Flag: EXEMPT BUGFIX
Change-Id: Ia6d75b51b086b84b1002b797fcc6a2e9a67b1017