APM: update device before route check if PolicyMix
When policy mix is active, it could be routing to devices suggested by
policy mix rather than which is suggested by the engine.

This discrepancy led to checking the unintended pair of ports and
invalidate such routes.

In this CL, the device is updated before checking the route, so that the
checks are applied on intended ports.

Bug: 462221662
Test: m
Flag: EXEMPT bugfix
Change-Id: I0b1079f9ddb12b1f0439ae8087ee7c45568aeb03