TagMonitor: Fix shorthand expansion
When TagMonitor parses tags, it looks for the "3a" shorthand and
attempts to do a substitution of the string in place.

The issue with the current code is, after looking for the first comma
after "3a" (getting the index `end`), it replaces "3a" in the string
*before* getting the rest of the string. As a consequence, when we try
to get the rest of the tag names after the comma, large portions (if not
all) would've been overwritten by the earlier "3a" replacement.

Bug: 469249014
Test: Tested with watch on caiman
Flag: EXEMPT BUGFIX
Change-Id: I212287ef4274d2f26fb5eee58303a2cac5b780f3